+++ src/client/client.h
@@ -360,8 +360,8 @@ void IN_Accumulate(void);
 /* cl_sequence.c (avi stuff) */
 qboolean CL_OpenAVIForWriting(char *filename);
 void CL_TakeVideoFrame(void);
-void CL_WriteAVIVideoFrame(const byte * imageBuffer, int size);
-void CL_WriteAVIAudioFrame(const byte * pcmBuffer, int size);
+void CL_WriteAVIVideoFrame(const byte * imageBuffer, size_t size);
+void CL_WriteAVIAudioFrame(const byte * pcmBuffer, size_t size);
 qboolean CL_CloseAVI(void);
 qboolean CL_VideoRecording(void);
 void CL_StopVideo_f(void);
